+++ glean-core/android/src/main/java/mozilla/telemetry/glean/private/CustomDistributionMetricType.kt
@@ -51,10 +51,4 @@ class CustomDistributionMetricType(
     @JvmOverloads
     fun testGetNumRecordedErrors(error: ErrorType, pingName: String? = null) =
         inner.testGetNumRecordedErrors(error, pingName)
-
-    @VisibleForTesting(otherwise = VisibleForTesting.NONE)
-    @JvmOverloads
-    fun testHasValue(pingName: String? = null): Boolean {
-        return this.testGetValue(pingName) != null
-    }
 }