+++ glean-core/android/src/main/java/mozilla/telemetry/glean/private/TimingDistributionMetricType.kt
@@ -64,10 +64,4 @@ class TimingDistributionMetricType(meta: CommonMetricData, timeUnit: TimeUnit) :
     @JvmOverloads
     fun testGetNumRecordedErrors(error: ErrorType, pingName: String? = null) =
         inner.testGetNumRecordedErrors(error, pingName)
-
-    @VisibleForTesting(otherwise = VisibleForTesting.NONE)
-    @JvmOverloads
-    fun testHasValue(pingName: String? = null): Boolean {
-        return this.testGetValue(pingName) != null
-    }
 }