+++ glean-core/android/src/test/java/mozilla/telemetry/glean/private/BooleanMetricTypeTest.kt
@@ -42,12 +42,10 @@ class BooleanMetricTypeTest {
         // Record two booleans of the same type, with a little delay.
         booleanMetric.set(true)
         // Check that data was properly recorded.
-        assertTrue(booleanMetric.testHasValue())
         assertTrue(booleanMetric.testGetValue()!!)
 
         booleanMetric.set(false)
         // Check that data was properly recorded.
-        assertTrue(booleanMetric.testHasValue())
         assertFalse(booleanMetric.testGetValue()!!)
     }
 
@@ -68,7 +66,7 @@ class BooleanMetricTypeTest {
         // Attempt to store the boolean.
         booleanMetric.set(true)
         // Check that nothing was recorded.
-        assertFalse(booleanMetric.testHasValue())
+        assertNull(booleanMetric.testGetValue())
     }
 
     @Test
@@ -103,12 +101,10 @@ class BooleanMetricTypeTest {
         // Record two booleans of the same type, with a little delay.
         booleanMetric.set(true)
         // Check that data was properly recorded in the second ping
-        assertTrue(booleanMetric.testHasValue("store2"))
         assertTrue(booleanMetric.testGetValue("store2")!!)
 
         booleanMetric.set(false)
         // Check that data was properly recorded in the second ping.
-        assertTrue(booleanMetric.testHasValue("store2"))
         assertFalse(booleanMetric.testGetValue("store2")!!)
     }
 }